Test CLI script to notebook
CLI inside notebook file, this allows for testing directly on the
python module responsible for the rendering of the notebooks.
