+++ bin/sphx_glr_python_to_jupyter.py
@@ -12,29 +12,9 @@ scripts into Jupyter Notebooks.
 # License: 3-clause BSD
 
 from __future__ import division, absolute_import, print_function
-import argparse
-import os
 
-import sphinx_gallery.gen_rst as sg
-from sphinx_gallery.notebook import jupyter_notebook, save_notebook
-
-parser = argparse.ArgumentParser(
-    description='Sphinx-Gallery Notebook converter')
-parser.add_argument('python_src_file', nargs='+',
-                    help='Input Python file script to convert. '
-                    'Supports multiple files and shell wildcards'
-                    ' (e.g. *.py)')
-
-
-def main():
-    args = parser.parse_args()
-
-    for src_file in args.python_src_file:
-        blocks = sg.split_code_and_text_blocks(src_file)
-        print('Converting {0}'.format(src_file))
-        example_nb = jupyter_notebook(blocks)
-        save_notebook(example_nb, src_file.replace('.py', '.ipynb'))
+from sphinx_gallery.notebook import python_to_jupyter_cli
 
 
 if __name__ == '__main__':
-    main()
+    python_to_jupyter_cli()
